Move RequestToken and AccessToken into Carto namespace
