+++ spec/models/user_part_crud_spec.rb
@@ -243,7 +243,7 @@ describe User do
           user = create_user(email: 'clientapp@example.com', username: 'clientapp', password: @user_password)
 
           user.create_client_application
-          user.client_application.access_tokens << ::AccessToken.new(
+          user.client_application.access_tokens << Carto::AccessToken.new(
             token: "access_token",
             secret: "access_secret",
             callback_url: "http://callback2",
@@ -252,7 +252,7 @@ describe User do
             client_application_id: user.client_application.id
           ).save
 
-          user.client_application.oauth_tokens << ::Carto::OauthToken.create!(
+          user.client_application.oauth_tokens << Carto::OauthToken.create!(
             token: "oauth_token",
             secret: "oauth_secret",
             callback_url: "http//callback.com",
@@ -272,7 +272,7 @@ describe User do
           user.destroy
 
           expect(ClientApplication.where(user_id: user.id).first).to be_nil
-          expect(AccessToken.where(user_id: user.id).first).to be_nil
+          expect(Carto::AccessToken.where(user_id: user.id).first).to be_nil
           expect(Carto::OauthToken.where(user_id: user.id).first).to be_nil
           $api_credentials.keys.should_not include(base_key)
         end
