+++ spec/support/factories/access_tokens.rb
@@ -2,7 +2,7 @@ module CartoDB
   module Factories
     def new_access_token(attributes = {})
       attributes = attributes.dup
-      AccessToken.new(attributes)
+      Carto::AccessToken.new(attributes)
     end
 
     def create_access_token(attributes = {})
